POS UI extension: target-agnostic intercepts capability declaration
Structural validation only; Core is the sole authority on which targets
support intercepts. Future-proof declaration shape on capabilities.intercepts.
